+++ src/index.js
@@ -2,14 +2,14 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './styles.css';
 import { BrowserRouter } from 'react-router-dom';
-import MathMagiciansApp from './MathMagiciansApp';
+import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 
 root.render(
   <React.StrictMode>
     <BrowserRouter>
-      <MathMagiciansApp />
+      <App />
     </BrowserRouter>
   </React.StrictMode>,
 );
